+++ src/lib/components/hooks/useDocumentContext.tsx
@@ -76,6 +76,9 @@ interface IDocumentContext {
   saveLayer: () => void
   groupId: string
   setGroupId: (groupId: string) => void
+  editGroupId: string
+  setEditGroupId: (groupId: string) => void
+  editLayer: { id: string; svg: string } | null
   layer: { id: string; svg: string } | null
   setLayer: (layer: { id: string; svg: string } | null) => void
   editStage: EDIT_STAGES