+++ CHANGELOG.md
@@ -3,8 +3,12 @@
 ## 0.5.5 : TBD
 
 - **Added**: Removing border from svg before saving/editing
+- **Added**: Layer in reading mode
+- **Added**: Button to side bar for choosing groups
+- **Added**: Help for new button in sidebar
 - **Changed**: Saving layer function
 - **Changed**: SVG in saving layer function from HTMLElement to string
+- **Fixed**: talwind in loader and helper
 
 ## 0.5.4 : 2024-07-17